docs(Intro): animation and page
